Implemented cpp file

+++ src/Sensors/Base/GPS.h
@@ -9,6 +9,7 @@
 #define __GPS_H__
 
 #include "Sensor.h"
+#include <cstdint>
 
 
 class GPS : public Sensor
@@ -23,7 +24,7 @@ public:
 
     virtual double getLatitude_deg() = 0;
     virtual double getLongitude_deg() = 0;
-    virtual uint16_t getSattelitesAmt() = 0;
+    virtual uint32_t getSattelitesAmt() = 0;
     virtual float getSpeed_kmph() = 0;
 };
